fix: Move misplaced mocks (#3500)
